+++ src/main/java/me/soo/helloworld/service/UserService.java
@@ -1,10 +1,37 @@
 package me.soo.helloworld.service;
 
+import lombok.RequiredArgsConstructor;
 import me.soo.helloworld.model.User;
+import me.soo.helloworld.repository.UserRepository;
+import me.soo.helloworld.util.PasswordEncoder;
+import org.springframework.stereotype.Service;
 
-public interface UserService {
+@Service
+@RequiredArgsConstructor
+public class UserService {
 
-    public void userSignUp(User user);
+    private final UserRepository userRepository;
 
-    public boolean isUserIdDuplicate(String userId);
+    private final PasswordEncoder passwordEncoder;
+
+    public void userSignUp(User user) {
+        String encodedPassword = passwordEncoder.encode(user.getPassword());
+        User encodedUser = User.builder()
+                .userId(user.getUserId())
+                .password(encodedPassword)
+                .email(user.getEmail())
+                .gender(user.getGender())
+                .birthday(user.getBirthday())
+                .originCountry(user.getOriginCountry())
+                .livingCountry(user.getLivingCountry())
+                .livingTown(user.getLivingTown())
+                .aboutMe(user.getAboutMe())
+                .build();
+
+        userRepository.insertUser(encodedUser);
+    }
+
+    public boolean isUserIdDuplicate(String userId) {
+        return userRepository.isUserIdDuplicate(userId);
+    }
 }
